+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Modified
+- Use '#[naked]' instread of global assembly in newer Rust version for RustSBI platforms
+
 ## [0.1.0] - 2020-12-26
 RustSBI is adapted to SBI standard with implementation number 4.
 ### Added